Version 1.7.0
##### 11/09/2023 Updates:

__Feature added: Show Related button - Visual Trace page__

Uses the selected message in the visual trace tab/FullTrace tab as a start location
Highlights messages in red that are directly related via the start or end icons on the SVG trace window
Hides messages in the FullTrace tab that are not related to this message
*Note this is a best estimate based on the start and end position of each message. For simple traces this will work fine, however it may fall down on more complicated traces where multiple messages are triggered from the same processor to the same operation and responses are sent back.


__Feature Added: Ctrl+Click message__

Ctrl clicking on a message's background in the FullTrace tab will scroll that message into view in the visual trace window


__General Visual Trace page features added:__

Messages in the FullTrace tab are now linked to their SVG equivalent in the Visual Trace frame. Click either and it clicks its partner. However over one and it and its partner will change colour slightly.
This is to allow you to more easily see which message is which when there are dozens of messages. For example, an ADT feed.


__Feature Added: Searchable Schema (individual segments & all)__

Expanding an individual segment to view its schema will now include a search bar. Type into this search bar to hide schema rows that don't include your typed text.
Use the button bar's Search Schema button to expand all currently available schema and replace the Search Schema button with a Search Bar. Clear the search bar and press ENTER to close this Search Bar and the expanded schema.
The individual segment's search bar won't appear when you click the button bar Search Schema button unless it was already open for that segment when you clicked Search Schema or you close and open an individual segment's Schema. Only the latest search made will apply to any segment with an individual search bar. When closing and reopening searches, no search is saved.


__Bugfix__: Blue.css file .css extension missing. Not sure where it went but it's back now and I'll do my best to keep it here.
__Bugfix__: Empty segments break the schema expansion. Now they will have to find another way to ruin your day.
__BugFix__: New Search History item appears when deleting one. Sometimes a search would rise from the ashes when deleting an old search. This Phoenix has been dealt with by the appropriate wizards.
__BugFix__: I thought changing the way buttons worked would be a good idea. It wasn't.
__BugFix__: Searches without criteria break the Search History. Now, if you feel so inclined, you can search without criteria and it should still function as intended.