Version 1.8.0
__Feature Updated__:

Stored instances export now includes all current settings. You can still import your old instances object, it will be immediately updated to the new version on import. Save this somewhere safe! This was probably more work than it was worth.

__Behaviour Change__: Bookmark folder is not automatically created. Now gets created when editing the settings according to the name chosen in the new Bookmark Folder Name setting. Yes, it is very fancy now.

__Bugfix__: Rule editor in Iris does not have coloured header. This now highlights. Excellent. 

__Feature Added__: ChatGPT Powered Test Message Generator.

When sending a test message on a production page you can now use ChatGPT to generate you a HL7 message. The prompt will appear as long as you have input an API Key to the ChatGPT API key setting of the Options page.

You need to select a schema and a doc type and it will automatically generate you a prompt. Edit the prompt as you please and then click the Generate to call ChatGPT. 


__Feature Added__: On Schema search you can now press enter to highlight matching copy. 

Highlighting the text on keypress proved too slow when several messages were open and highlighting on enter seemed an acceptable compromise. Enter remains the exit key when the search bar is blank. 

When searching individual segments text is automatically highlighted. Segment search highlighting takes priority sometimes, depends how it's feeling.


__Feature Added__: Active Production Category Search

On productions a button appears next to the Category dropdown. Clicking this button loads active productions onto the page, gathers their category lists and complies them into a dropdown. Choosing a category from this dropdown will cause all productions with matching categories to show on the page with this Category filter applies to them. Any categories on the starting production page are prefixed with *. This is for ease of identification. Choosing a category that is not on the starting production page will not update the starting production page.

Note it only shows ACTIVE productions. That is, productions shown as Running on your %SYS homepage.

This is useful for seeing workflows for a particular category across namespaces. It's a little slow if you have a lot of productions, but it works pretty well. 


__Feature Added__: Segment Search.

Hide segments that don't match your current search criteria. You can search specific segments by prepending your search with \SEGMENT\YOUR SEARCH HERE
e.g. 
\OBX\151570

Bonus points if you know which Rosetta Terminology vitals that references.

That example will only hide OBX segments, leaving other segments still visible. Useful for searching for specific OBX-3 values but leaving the OBR/PID/MSH etc.. segments showing. 

Please remember to use back slashes. This also renders back slashes useless for search purposes. Sorry.


__Feature Added__: Hide Iris Whiz Buttons Until Needed

Added an IRIS WHIZ button to maximise/minimise the button bar as needed. This is the bar that appears on Full Trace tab, Message Contents Tab and Selected Messages Tab and contains Text Compare, Schema Search etc.

Choose on the option page whether you want the button bar to be expanded on page load. Default is buttons are hidden. It also looks more fancy now. You're welcome.


__Feature Added__: Bookmark Folder Name setting

A new text input in the settings adds ability to recreate the bookmark folder with your desired name. I've not auto-deleted the old one in case people add other bookmarks to this folder. You have been warned.


__Feature added__: Schema Mode on Message Contents pages.

You can now toggle Schema Mode using the… Schema Mode button! The current behaviour for expanding schema only shows the known schema that is available in the Title of the field - the bit where you hover your mouse over the fields. The new version can look up the schema using the link in the segment element. This is slow, especially when used in conjunction with the Searching features. Therefore I've limited it to the Contents view. WARNING! It will continue to work if you import messages to this view. Unless you like to watch the world burn, don't use the Full Schema option if you have many messages imported into your page. 


__Update__: Minor speed increases. Nobody will notice. Yay.